requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
dnspython==2.4.2
bcrypt==4.1.2
PyJWT==2.8.0
python-multipart==0.0.6
